Can not get sub test module imports working immediately
