+++ src/tests/test_plaform_source_base.py
@@ -2,7 +2,6 @@ import pytest
 import requests_mock
 
 from unittest import TestCase
-from unittest.mock import MagicMock, Mock, patch
 
 from platformen.scraper import PlatformSourceConfig, PlatformSource, ScrapeException
